Extract all pipeline steps to separate files
